feat(docz-core): add two new plugin hooks
This change adds two new plugin hooks to docz-core supporting
functionality previously not supported.

It adds plugin.modifyFiles which gives plugins the chance to edit the
list of mdx source files before docz starts processing them.

It also adds plugin.setConfigPromise which is the async analogue of
plugin.setConfig. This allows plugins to run more dynamic setup
processes such as modifying the docz config based on the project's
filesystem state or a sub-build process.

Note that the naming of setConfig vs setConfigPromise is following the
precedent set by Webpack's plugin hooks which is explained here:

https://webpack.js.org/api/plugins/#plugin-types